+++ src/frontend/ui/commun/Header/Header.tsx
@@ -11,7 +11,7 @@ import "@gouvfr/dsfr/dist/component/link/link.min.css";
 import "@gouvfr/dsfr/dist/component/modal/modal.min.css";
 import styles from "./Header.module.css";
 import { Role } from "../../../../commons/Role";
-import { CONSOLE_ADMIN } from "../../../utils/nomenclature-matomo";
+import { CONSOLE_ADMIN, RECHERCHE_SIMPLE } from "../../../utils/nomenclature-matomo";
 import { useFavoris } from "../../favoris/useFavoris";
 import { Breadcrumb } from "../Breadcrumb/Breadcrumb";
 import { BtnRetourRecherche } from "../BtnRetourRecherche/BtnRetourRecherche";
@@ -182,6 +182,7 @@ export const Header = () => {
                         <button
                           className="fr-btn"
                           onClick={(event) => {
+                            sendEvent(RECHERCHE_SIMPLE);
                             event.preventDefault();
                             localStorage.setItem("searchItem", encodeURIComponent(terme));
                             router.push(paths.ACCUEIL + "?terme=" + encodeURIComponent(terme), paths.ACCUEIL);