+++ src/frontend/utils/nomenclature-matomo.ts
@@ -11,3 +11,7 @@ export const EXPORT = (source: string) => {
 // Navigation
 export const CONSOLE_ADMIN = {category: 'navigation', action: 'console_admin'};
 export const CARTOGRAPHIE = {category: 'navigation', action: 'cartographie'};
+
+// Recherche 
+export const RECHERCHE_SIMPLE = {category: 'recherche', action: 'simple'};
+