+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 ### Bug Fixes
 
 ### Features
-hel-1050: Ajout d’event Matomo lors de l’export
-hel-1057: Ajout d'event Matomo lors de la consultation de la console admin
-hel-1054: Ajout d'event Matomo lors de la consultation de la cartographie
+- hel-1050: Ajout d’event Matomo lors de l’export
+- hel-1057: Ajout d'event Matomo lors de la consultation de la console admin
+- hel-1054: Ajout d'event Matomo lors de la consultation de la cartographie
+- hel-1053: Ajout d'event Matomo lors de la recherche simple
 
 ## [1.8.10](https://github.com/DNUM-SocialGouv/helios/compare/release/1.8.9...release/1.8.10) (2026-06-15)